refactor(NewsletterDialog): consolidate, port to ts
