+++ client/src/dialogs/About/AboutDialog.scss
@@ -74,7 +74,8 @@
   > :first-child {
     margin-top: 0;
   }
-  x > h2 {
+
+  > h2 {
     margin-top: 1.5em;
   }
 }
