+++ client/src/dialogs/DialogRoot.jsx
@@ -11,7 +11,7 @@ import SaveAsImageDialog from './SaveAsImage'
 import SettingsDialog from './Settings'
 import SignInDialog from './SignInDialog'
 import WhatsNewDialog from './WhatsNewDialog'
-import NewsletterDialog from './NewsletterDialog'
+import NewsletterDialog from './Newsletter'
 import UpgradeDialog from './UpgradeDialog'
 import SentimentSurveyDialog from './SentimentSurveyDialog'
 import ErrorDialog from './ErrorDialog'
