+++ client/src/dialogs/Newsletter/NewsletterDialog.scss
@@ -1,5 +1,4 @@
 .newsletter-dialog {
-  min-width: 400px;
   max-width: 520px;
 }
