+++ client/src/dialogs/Newsletter/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NewsletterDialog'
